feat: genetic algorithm overhaul (#27)
## Description

This PR aims to overhaul current implementation of genetic algorithm. 

Major pain points in current implementation:

* It is not
[pipeline](https://medium.com/@bonnotguillaume/software-architecture-the-pipeline-design-pattern-from-zero-to-hero-b5c43d8a4e60)
oriented as it should be (idea coming from discussions with client).
Besides that, current algorithm configuration style is not elegant.
* It lacks flexibility: currently only `f64` vectors are supported as
chromosomes.
* Many more reasons... (TODO)

This PR added builder pattern for construction of GA instance. Pipeline
& support for other chromosome representation in next PRs.

## Linked issues

It partially deals with #21 but it does not introduce a pipeline as for
now.

## Important implementation details

TODO