feat: add first type of rank selection  (#71)
## Description

Should be merged only after

* #69 

is merged.

Added missing selection operator. 

This is one of two popular implementations. Second one has additional
`r` parameter, which is used to decide which of random selected
individuals to use as a parent. I left the second implementation behind
for now, as presence of additional parameter raises some problems with
our current approach to operator implementations. Namely we can't just
add additional param to function declaration as it would break existing
selection operator API in GA implementation (how do we decide whether to
pass 2 or 3 params to operator in runtime?). One possible solution might
be to implement operators as traits (concrete implementations as structs
with `FnMut` trait implemented). Such approach would not only allow for
operator parameterization but also for some kind of function overload
which is not present in Rust.

## Linked issues

Closes #37 

## Important implementation details

N/A