feat(GA): `replacement` module & `BothParent` operator (#263)
## Description

First thing to note is that current solution is not really a "steady"
replacement.

Currently I've introduced a separate `replacement` operator with the
same approach as for other operators: trait + concrete types that
implement that trait.

What makes it "not steady" is that this operator is passed to GA in the
same way as any other operator, and thus it is called separately, after
the crossover and there is need to iterate over whole population once
again...

To make it "steady" we could pass replacement operator directly to the
crossover operator as a parameter / construction parameter - and this is
the thing I want your opinion on the most. **Should we make the
replacement a part of crossover process or should it be a standalone
operator?**

The second topic is that after `ChromosomeWrapper` was removed in
https://github.com/ecrs-org/ecrs/pull/147 there is no option to define
custom state on a individual, thus it makes some variants of replacement
operator basically impossible to implement in a reasonable way (such as
oldest individual replacement - as you would need to have a lifetime
associated with the individual). **Therefore I want to open a discussion
whether we should bring back `ChromosomeWrapper` (or some improved
version of this concept)**.

If I did not explain things clearly, please ask. 

## Linked issues

Resolves #58

## Important implementation details

Described in previous sections.