feat(ga): add `generate_iter` method on `PopulationGenerator` trait (#404)
## Description

Add `generate_iter` method on `PopulationGenerator` trait.

Default implementation is provided, based on `generate` method -- I know
that it makes little sense, because vector has to be allocated anyway,
but I wanted to avoid refactoring of all operators for now.

TODO: Refactor `PopulationGenerator` trait so that `generate` method is
based on `generate_iter`.

## Linked issues

Closes #403