perf(jssp)!?: reduce complexity of `delay_feasibles` computation (#451)
<!-- If applicable - remember to add the PR to the EA Rust project (ONLY
IF THERE IS NO LINKED ISSUE) -->

## Description

I've used guarantee introduced in #444 

We do not need to iterate over all predecessors when checking whether
all predecessors will have been finished in before `t_g + delay_g`. It
is sufficient to check only the direct one, because it could have been
scheduled only in case its own direct predecessor had finished (and so
on...). However we need to handle special case of sink operation as it
has every every operation in it's predecessor list. We do not need to
handle operation zero, as it is always scheduled up front, before first
call to this method ==> it is filtered out by previous predicate.