feat!: pass GAMetadata to all operator calls (#468)
## Description

Breaking change.

This PR introduces passing `&GAMetadata` to all operator calls.

In the future either metadata could be extended to contain more
information
or we would create new structure `OperatorCallContext` which would
contain more information (e.g. population statistics),
so that operators can base their decisions on it.

## Linked issues <!-- Please use "Resolves #<issue_no> syntax in case
this PR should be linked to an issue -->

Closes #467