fix: add important comment and update `total_dur` just before `on_end` callback (#472)
<!-- If applicable - remember to add the PR to the EA Rust project (ONLY
IF THERE IS NO LINKED ISSUE) -->

## Description

Decided not to update `total_dur` before each hook call **for now**, to
avoid vastly increasing number of syscalls.
This, however should be reconsidered in the future.

Added important comment to **slightly** decrease likelihood of future
misuse of this field.

## Linked issues <!-- Please use "Resolves #<issue_no> syntax in case
this PR should be linked to an issue -->

Closes #471

## Important implementation details <!-- if any, optional section -->